# internal/openapi/openapi.json
{
  "components": {
    "responses": {
...
      "BalanceInvalid": {
        "content": {
          "application/json": {
            "examples": {
              "invalid_balance": {
                "value": {
                  "error": {
                    "code": "invalid_balance"
                  }
                }
              },
              "invalid_request": {
                "value": {
                  "error": {
                    "code": "invalid_request"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The balance request, resource key, or idempotency key is invalid."
      },
...
    },
    "schemas": {
...
      "AddBalanceRequest": {
        "additionalProperties": false,
        "properties": {
          "customerId": {
            "example": "user_123",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "meterKey": {
            "example": "api_calls",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "example": 10000,
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "customerId",
          "meterKey",
          "quantity"
        ],
        "type": "object"
      },
...
      "Balance": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customerId": {
            "maxLength": 255,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "meterKey": {
            "maxLength": 120,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "meterKey",
          "quantity",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "localCookieSession": {
        "in": "cookie",
        "name": "consumel_session",
        "type": "apiKey"
      },
      "productionCookieSession": {
        "in": "cookie",
        "name": "__Host-consumel_session",
        "type": "apiKey"
      },
      "projectAPIKey": {
        "bearerFormat": "cm_test_… or cm_live_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Consumel usage-based billing infrastructure API.",
    "title": "Consumel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/balances": {
      "post": {
        "parameters": [
          {
            "description": "A UUID v7 that identifies this logical balance addition. Reuse it only when retrying the same request.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "customerId": "user_123",
                "meterKey": "api_calls",
                "quantity": 10000
              },
              "schema": {
                "$ref": "#/components/schemas/AddBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                  "meterKey": "api_calls",
                  "quantity": 10000,
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            },
            "description": "Completed successfully.",
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present with the value true when Consumel returns the original result of a committed retry.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/BalanceSubjectNotFound"
          },
          "409": {
            "$ref": "#/components/responses/BalanceConflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Add an idempotent quantity to a customer and meter balance.",
        "tags": [
          "Balances"
        ]
      }
    },
    "/v1/balances/{customerId}/{meterKey}": {
      "put": {
        "parameters": [
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "description": "The stable meter key supplied when the meter was created.",
            "in": "path",
            "name": "meterKey",
            "required": true,
            "schema": {
              "maxLength": 120,
              "pattern": "^[a-z][a-z0-9_-]*$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "quantity": 10000
              },
              "schema": {
                "$ref": "#/components/schemas/SetBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                  "meterKey": "api_calls",
                  "quantity": 10000,
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/BalanceSubjectNotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Set a customer and meter balance to an exact quantity.",
        "tags": [
          "Balances"
        ]
      }
    },
...
    "/v1/customers": {
      "get": {
        "parameters": [
          {
            "description": "A case-insensitive customer ID, name, or email search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of customers to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "customers": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "customerId": "user_123",
                      "email": "jordan@example.com",
                      "metadata": {
                        "country": "US",
                        "location": "New York, NY",
                        "plan": "growth"
                      },
                      "name": "Jordan Lee",
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/Customers"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "List customers in the API key's project environment.",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "customerId": "user_123",
                "email": "jordan@example.com",
                "metadata": {
                  "country": "US",
                  "location": "New York, NY",
                  "plan": "growth"
                },
                "name": "Jordan Lee"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "409": {
            "$ref": "#/components/responses/CustomerConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Create a customer in the API key's project environment.",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/v1/customers/{customerId}/balances": {
      "get": {
        "parameters": [
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "balances": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "customerId": "user_123",
                      "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                      "meterKey": "api_calls",
                      "quantity": 10000,
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/Balances"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/BalanceNotFound"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "List a customer's active meter balances in the API key's project environment.",
        "tags": [
          "Balances"
        ]
      }
    },
    "/v1/customers/{customerId}/balances/{meterKey}": {
      "get": {
        "parameters": [
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "description": "The stable meter key supplied when the meter was created.",
            "in": "path",
            "name": "meterKey",
            "required": true,
            "schema": {
              "maxLength": 120,
              "pattern": "^[a-z][a-z0-9_-]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                  "meterKey": "api_calls",
                  "quantity": 10000,
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/BalanceNotFound"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Return one customer and meter balance from the API key's project environment.",
        "tags": [
          "Balances"
        ]
      }
    },
...
    "/v1/projects/{projectId}/environments/{environment}/balances": {
      "post": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "A UUID v7 that identifies this logical balance addition. Reuse it only when retrying the same request.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "customerId": "user_123",
                "meterKey": "api_calls",
                "quantity": 10000
              },
              "schema": {
                "$ref": "#/components/schemas/AddBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                  "meterKey": "api_calls",
                  "quantity": 10000,
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            },
            "description": "Completed successfully.",
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present with the value true when Consumel returns the original result of a committed retry.",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/BalanceSubjectOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/BalanceOrEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Add an idempotent quantity to a customer and meter balance from the signed-in project workspace.",
        "tags": [
          "Dashboard Balances"
        ]
      }
    },
    "/v1/projects/{projectId}/environments/{environment}/balances/{customerId}/{meterKey}": {
      "put": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "description": "The stable meter key supplied when the meter was created.",
            "in": "path",
            "name": "meterKey",
            "required": true,
            "schema": {
              "maxLength": 120,
              "pattern": "^[a-z][a-z0-9_-]*$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "quantity": 10000
              },
              "schema": {
                "$ref": "#/components/schemas/SetBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                  "meterKey": "api_calls",
                  "quantity": 10000,
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/BalanceSubjectOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/BalanceEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Set a customer and meter balance to an exact quantity from the signed-in project workspace.",
        "tags": [
          "Dashboard Balances"
        ]
      }
    },
...
    "/v1/projects/{projectId}/environments/{environment}/customers/{customerId}/balances": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "balances": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "customerId": "user_123",
                      "id": "0199aa81-ce8c-73bf-a880-8e84654b9a6c",
                      "meterKey": "api_calls",
                      "quantity": 10000,
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/Balances"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/BalanceInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/BalanceOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/BalanceEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/BalanceFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "List a customer's active meter balances from the signed-in project workspace.",
        "tags": [
          "Dashboard Balances"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.consumel.com"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8080"
    }
  ]
}